\name{data_dposs}
\alias{data_dposs}
\docType{data}
\title{
DPOSS dataset
}
\description{
This is a random subset of 20'000 stars from the Digitized 
Palomar Sky Survey (DPOSS) described by Odewahn et al. (1998).
}

\usage{data("data_dposs")}

\format{
  A matrix of dimensions \eqn{20000 \times 21}.
}

\references{
Odewahn, S., S. Djorgovski, R. Brunner, and R. Gal (1998). Data From the Digitized Palomar Sky Survey. Technical report, California Institute of Technology.
}
\examples{
data("data_dposs")
# For more examples, we refer to the vignette:
vignette("MacroPCA_examples")
}
