\name{data_glass}
\alias{data_glass}
\docType{data}
\title{
The glass dataset
}
\description{
A dataset containing spectra with \eqn{d = 750} wavelengths collected on
\eqn{n = 180} archeological glass samples.
}
\usage{data("data_glass")}
\format{
  A data frame with 180 observations of 750 wavelengths.

}

\source{
Lemberge, P., De Raedt, I., Janssens, K.H., Wei, F., and Van Espen, P.J. (2000).
Quantitative Z-analysis of 16th-17th century archaeological glass vessels using
PLS regression of EPXMA and \eqn{\mu}-XRF data. \emph{Journal of Chemometrics}, \bold{14},
751--763.
}

\examples{
data("data_glass")
}
