\name{data_mortality}
\alias{data_mortality}
\docType{data}
\title{
The mortality dataset
}
\description{
This dataset contains the mortality by age for males in France, from 1816 to 2013
as obtained from the Human Mortality Database.
}
\usage{data("data_mortality")}
\format{
  A data frame with 198 calendar years (rows) and 91 age brackets (columns).
 
}

\source{
Human Mortality Database. University of California, Berkeley (USA), and
Max Planck Institute for Demographic Research (Germany). Available at
 \url{https://www.mortality.org} (data downloaded in November 2015).
}
\references{
Hyndman, R.J., and Shang, H.L. (2010), Rainbow plots, bagplots, and boxplots for
functional data, \emph{Journal of Computational and Graphical Statistics}, \bold{19}, 29--45.
}
\examples{
data("data_mortality")
}

